benchmarks/requirements.txt
requests==2.31.0
openai==1.12.0
anthropic==0.18.1
tiktoken==0.6.0
python-dotenv==1.0.1
tqdm==4.66.2
pandas==2.2.0
numpy==1.26.4
